requirements.txt
requests
bunnycdnpython
niconico.py
fastapi
python-dotenv